Register prims as operators.
This makes prims look as if they were defined in native_functions.yaml
but they're still all written in Python.  You now need to give a full
schema string for your prims.  The returned prim object is now
torch.ops.prim overload (prims are not allowed to be overloaded,
so we return the overload, not the overload packet, for speed.)

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 24495013ecd4a30b1469aa1983df359168898ba6
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77117