[JIT] Support for ParameterDict getattr
Adds support for scripting ParameterDicts and getattr() on them. It does
not support iterating on ParameterDicts because torch/nn/container.py
implementation of ParameterDict.items() uses a generator, which is not
supported by torchscript. torch/nn/container.py would need to be updated
so that iter gets correctly registered in python_sugared_value.cpp

Added a test in test_module_containers.py

ghstack-source-id: 0aab18d60b2d66229878902e382207a2f0789b0a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77143